Replace floats with doubles

+++ src/global.h
@@ -151,7 +151,7 @@ struct Block {
   Block         *cnr[2]; // adjacent block in contiguous set
   ContiguousSet *cset;   // contiguous set id
   long          pos[2];  // start and stop positions
-  float         score;   // score provided by synteny program
+  double         score;   // score provided by synteny program
   size_t        grpid;   // overlapping group id;
   char          strand;  // strand [+-.]
   size_t        linkid;  // a unique block id used mostly for debugging
